RDKAAMP-2206 address 2402 compilation warnings
Reason for change: tech debt
Test Procedure: regression only (gst test harness)
Risks: Low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:dev_sprint_24_1
DEPS_BRANCH=rdk/components/generic/aampabr:dev_sprint_24_1
DEPS_BRANCH=rdk/components/generic/aampmetrics:dev_sprint_24_1

+++ priv_aamp.cpp
@@ -12393,7 +12393,7 @@ void PrivateInstanceAAMP::ProcessID3Metadata(char *segment, size_t size, MediaTy
 /**
  * @brief Report ID3 metadata events
  */
-void PrivateInstanceAAMP::ReportID3Metadata(MediaType mediaType, const uint8_t* ptr, uint32_t len,
+void PrivateInstanceAAMP::ReportID3Metadata(MediaType mediaType, const uint8_t* ptr, size_t len,
 	const char* schemeIdURI, const char* id3Value, uint64_t presTime,
 	uint32_t id3ID, uint32_t eventDur, uint32_t tScale, uint64_t tStampOffset)
 {
